+++ mobile/android/android-components/components/support/ktx/src/main/java/mozilla/components/support/ktx/android/content/SharedPreferences.kt
@@ -67,11 +67,21 @@ private class LongPreference(
 
 private class StringPreference(
     private val key: String,
-    private val default: String
+    private val default: String,
+    private val persistDefaultIfNotExists: Boolean = false,
 ) : ReadWriteProperty<PreferencesHolder, String> {
 
-    override fun getValue(thisRef: PreferencesHolder, property: KProperty<*>): String =
-        thisRef.preferences.getString(key, default) ?: default
+    override fun getValue(thisRef: PreferencesHolder, property: KProperty<*>): String {
+        return thisRef.preferences.getString(key, null) ?: run {
+            when (persistDefaultIfNotExists) {
+                true -> {
+                    thisRef.preferences.edit().putString(key, default).apply()
+                    thisRef.preferences.getString(key, null)!!
+                }
+                false -> default
+            }
+        }
+    }
 
     override fun setValue(thisRef: PreferencesHolder, property: KProperty<*>, value: String) =
         thisRef.preferences.edit().putString(key, value).apply()
@@ -147,17 +157,26 @@ fun longPreference(key: String, default: Long): ReadWriteProperty<PreferencesHol
 
 /**
  * Property delegate for getting and setting a string shared preference.
+ * Optionally this will persist the default value if one is not already persisted.
  *
  * Example usage:
  * ```
  * class Settings : PreferenceHolder {
  *     ...
- *     var permissionsEnabledEnum by stringPreference("permissions_enabled", default = "blocked")
+ *     var permissionsEnabledEnum by stringPreference(
+ *          "permissions_enabled",
+ *          default = "blocked",
+ *          persistDefaultIfNotExists = true,
+ *     )
  * }
  * ```
  */
-fun stringPreference(key: String, default: String): ReadWriteProperty<PreferencesHolder, String> =
-    StringPreference(key, default)
+fun stringPreference(
+    key: String,
+    default: String,
+    persistDefaultIfNotExists: Boolean = false,
+): ReadWriteProperty<PreferencesHolder, String> =
+    StringPreference(key, default, persistDefaultIfNotExists)
 
 /**
  * Property delegate for getting and setting a string set shared preference.